try to put things together #12, Ubicall/node-red-contrib-ivr#31
