+++ red/ubicall/plist/nodes/utils.js
@@ -8,6 +8,10 @@ function isZendeskNode(node) {
   )
 }
 
+function isZendeskKBNode(node) {
+  return ((node.hasOwnProperty('type') && node.type == 'view-zendesk-knowledge-base'))
+}
+
 function getZendeskTicketFormNodes(nodes) {
   return nodes.filter(isZendeskFormNode);
 }
@@ -15,6 +19,15 @@ function getZendeskTicketFormNodes(nodes) {
 function hasZendeskNodes(nodes) {
   for (var i = 0; i < nodes.length; i++) {
     if (isZendeskNode(nodes[i])) {
+      return true;
+    }
+  }
+  return false;
+}
+
+function hasZendeskKBNodes(nodes) {
+  for (var i = 0; i < nodes.length; i++) {
+    if (isZendeskKBNode(nodes[i])) {
       return true;
     }
   }
@@ -71,6 +84,8 @@ module.exports = {
   isZendeskNode: isZendeskNode,
   getZendeskTicketFormNodes: getZendeskTicketFormNodes,
   hasZendeskNodes: hasZendeskNodes,
+  hasZendeskKBNodes: hasZendeskKBNodes,
+  isZendeskKBNode: isZendeskKBNode,
   getNodeWithId: getNodeWithId,
   getStartNode: getStartNode,
   concat: _concat
