+++ test/red/ubicall/plist/3rd/zendesk/help-center_spec.js
@@ -1,0 +1,11 @@
+var should = require("should");
+var sinon = require("sinon");
+var util = require("util");
+
+var log = require("../../red/log");
+
+describe("ubicall/plist/", function() {
+  it('should load kb for ubicall', function() {
+    
+  });
+});
